impelement XgbResultDisplay
