--- xgbsearch/XgbSearchDisplay.py
@@ -1,8 +1,0 @@
-import pandas as pd
-import numpy as np
-
-
-class XgbSearchDisplay:
-
-    def __init__(self):
-        pass
