+++ xgbsearch/__init__.py
@@ -1,3 +1,3 @@
 from .XgbSearch import XgbSearch, XgbGridSearch, XgbRandomSearch
-from .XgbSearchDisplay import XgbSearchDisplay
 from .XgbSearchExampleData import get_titanic
+from .XgbResultDisplay import XgbResultDisplay
